Finish breaking donation out from branding bar
