+++ example/index.html
@@ -12,7 +12,8 @@
 </head>
 
 <body>
-
+    <!-- <iframe src="http://localhost:4000/modal/modal.html" frameborder="0" allowtransparency="true" style="z-index: 9999; display: block; border: 0px none transparent; overflow-x: hidden; overflow-y: auto; visibility: visible; margin: 0px; padding: 0px; -webkit-tap-highlight-color: transparent; position: fixed; left: 0px; top: 0px; width: 100%; height: 100%;"></iframe> -->
+    
     <!-- Branding Bar -->
     <div class="branding-bar" data-bb-brandingbar="true" data-bb-property-id="sunlightlabs-awesome"></div>
 
@@ -31,7 +32,7 @@
         </div>
     </div>
 
-    <script src="dist/js/brandingbar.min.js"></script>
+    <script src="dist/js/brandingbar.js"></script>
     <!--[if IE 8]>
         <script src="https://s3.amazonaws.com/sf-icons/js/sf-icons.js"></script>
     <![endif]-->
