+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "sunlight-brandingbar",
-  "version": "0.4.0",
+  "version": "0.4.1",
   "author": "Sunlight Foundation <contact@sunlightfoundation.com>",
   "repository": {
     "type": "git",
